app/etl-job/requirements.txt
google-cloud-storage
python-dotenv
google-cloud-bigquery